Criada app de usuário e formulários de registros

+++ agendamanha/diario/views.py
@@ -1,17 +1,26 @@
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import Http404
 from django.shortcuts import render, redirect
 
 from agendamanha.diario.models import Diario, DiarioForm
 
 
+def check_entrada_owner(entrada, request):
+    if entrada.owner != request.user:
+        raise Http404
+
+
+@login_required
 def diario(request):
     """Mostra todas as afirmações."""
-    entradas = Diario.objects.order_by('date_added')
+    entradas = Diario.objects.filter(owner=request.user).order_by('date_added')
     entradas = entradas.reverse()
     context = {'entradas': entradas}
     return render(request, 'diario/diario.html', context)
 
 
+@login_required
 def add_entrada(request):
     """Add nova entrada no diário"""
     if request.method != 'POST':
@@ -19,15 +28,20 @@ def add_entrada(request):
     else:
         form = DiarioForm(data=request.POST)
         if form.is_valid():
-            form.save()
+            nova_entrada = form.save(commit=False)
+            nova_entrada.owner = request.user
+            nova_entrada.save()
             return redirect('diario:diario')
 
     context = {'form': form}
     return render(request, 'diario/add_entrada.html', context)
 
 
+@login_required
 def edit_entrada(request, entrada_id):
     entrada = Diario.objects.get(id=entrada_id)
+
+    check_entrada_owner(entrada, request)
 
     if request.method != 'POST':
         form = DiarioForm(instance=entrada)
@@ -42,8 +56,12 @@ def edit_entrada(request, entrada_id):
     return render(request, 'diario/edit_entrada.html', context)
 
 
+@login_required
 def del_entrada(request, entrada_id):
     entrada = Diario.objects.get(id=entrada_id)
+
+    check_entrada_owner(entrada, request)
+
     entrada.delete()
 
     messages.info(request, 'Entrada do diário deletada com sucesso.')
